requirements.txt
aws-cdk.core==1.137.0
aws-cdk.assertions==1.137.0
aws-cdk.aws_eks==1.137.0
aws-cdk.aws_ecr_assets==1.137.0
aws_cdk.aws_ec2==1.137.0
aws_cdk.aws_codebuild==1.137.0
aws_cdk.aws_codepipeline==1.137.0
aws_cdk.aws_codepipeline_actions==1.137.0
aws_cdk.aws_codestarconnections==1.137.0
aws_cdk.aws_s3_assets==1.137.0